+++ src/idtrackerai/network/__init__.py
@@ -3,7 +3,7 @@
 # NetworkParams should be loaded before LearnerClassification
 from torch.backends import cudnn
 
-from .utils import normalize, fc_weights_reinit, weights_xavier_init, DEVICE
+from .utils import fc_weights_reinit, weights_xavier_init, DEVICE
 from .network_params import NetworkParams
 from .learners import LearnerClassification
 from .train import train, evaluate
@@ -15,7 +15,6 @@ __all__ = [
     "LearnerClassification",
     "train",
     "weights_xavier_init",
-    "normalize",
     "fc_weights_reinit",
     "NetworkParams",
     "DEVICE",