+++ docs/source/user_guide/changelog.rst
@@ -9,6 +9,7 @@ Authors since :ref:`5.0.0`: Jordi Torrents
 - Simplify video JSON file.
 - Add boolean flag ``identity_transfer_succeded`` to video JSON file.
 - Remove ``accumulation_statistics`` from video JSON file.
+- Faster training with efficient image normalization
 
 5.2.1
 =====